requirements.txt
Django==2.0.7
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
django-registration-redux==2.4
django-widget-tweaks==1.4.2
idna==2.7
python-decouple==3.1
pytz==2018.5
requests==2.19.1
urllib3==1.23
